requirements.txt
asn1crypto==1.5.1
boto3==1.40.67
botocore==1.40.67
greenlet==3.2.4
jmespath==1.0.1
numpy==2.2.6
pandas==2.3.2
pg8000==1.31.5
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
s3transfer==0.14.0
scramp==1.4.6
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
fastparquet==2024.11.0
fsspec==2025.10.0
cramjam==2.11.0
packaging==25.0